Move workflow models to the correct directory.
