+++ Lombiq.HelpfulExtensions/Extensions/ContentSets/Workflows/Handlers/WorkflowContentSetEventHandler.cs
@@ -2,6 +2,7 @@
 using Lombiq.HelpfulExtensions.Extensions.ContentSets.Models;
 using Lombiq.HelpfulExtensions.Extensions.ContentSets.ViewModels;
 using Lombiq.HelpfulExtensions.Extensions.ContentSets.Workflows.Activities;
+using Lombiq.HelpfulExtensions.Extensions.ContentSets.Workflows.Models;
 using Lombiq.HelpfulLibraries.OrchardCore.Workflow;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
